% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{AKer.epa}
\alias{AKer.epa}
\title{Asymmetric Epanechnikov Kernel}
\usage{
AKer.epa(u)
}
\arguments{
\item{u}{Numeric vector of evaluation points.}
}
\value{
Kernel values at u (0 outside [0, 1]).
}
\description{
Asymmetric Epanechnikov Kernel
}
\examples{
u <- seq(-0.5, 1.5, length.out = 100)
plot(u, AKer.epa(u), type = "l", main = "Asymmetric Epanechnikov Kernel")
}
