% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irregFdata.R
\name{autoplot.irregFdata}
\alias{autoplot.irregFdata}
\title{Autoplot method for irregFdata objects}
\usage{
\method{autoplot}{irregFdata}(object, ..., alpha = 0.7)
}
\arguments{
\item{object}{An irregFdata object.}

\item{...}{Additional arguments (ignored).}

\item{alpha}{Transparency for lines.}
}
\value{
A \code{ggplot} object.
}
\description{
Autoplot method for irregFdata objects
}
