% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster.init}
\alias{cluster.init}
\title{K-Means++ Center Initialization}
\usage{
cluster.init(fdataobj, ncl, metric = "L2", seed = NULL)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{ncl}{Number of clusters.}

\item{metric}{Metric to use. One of "L2", "L1", or "Linf".}

\item{seed}{Optional random seed.}
}
\value{
An fdata object containing the initial cluster centers.
}
\description{
Initialize cluster centers using the k-means++ algorithm, which selects
centers with probability proportional to squared distance from existing
centers.
}
\examples{
t <- seq(0, 1, length.out = 50)
X <- matrix(rnorm(30 * 50), 30, 50)
fd <- fdata(X, argvals = t)
init_centers <- cluster.init(fd, ncl = 3)
}
