% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{cov}
\alias{cov}
\title{Functional Covariance Function}
\usage{
cov(fdataobj, ...)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A list with components:
\describe{
\item{cov}{The covariance matrix (m x m for 1D, (m1\emph{m2) x (m1}m2) for 2D)}
\item{argvals}{The evaluation points (same as input)}
\item{mean}{The mean function}
}
}
\description{
Computes the covariance function (surface) for functional data.
For 1D: \code{Cov(s, t) = E[(X(s) - mu(s))(X(t) - mu(t))]}
For 2D: Covariance across the flattened domain.
}
\examples{
# 1D functional data
t <- seq(0, 1, length.out = 50)
X <- matrix(0, 20, 50)
for (i in 1:20) X[i, ] <- sin(2*pi*t) + rnorm(50, sd = 0.2)
fd <- fdata(X, argvals = t)
cov_result <- cov(fd)
image(cov_result$cov, main = "Covariance Surface")
}
