% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{depth.RT}
\alias{depth.RT}
\title{Random Tukey Depth}
\usage{
depth.RT(fdataobj, fdataori = NULL, ...)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{fdataori}{Reference sample (default: fdataobj itself).}

\item{...}{Additional arguments.}
}
\value{
Numeric vector of depth values.
}
\description{
Wrapper for \code{depth(method = "RT")}. Useful as a function argument.
}
\seealso{
\code{\link{depth}}
}
