% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdars-package.R
\docType{package}
\name{fdars-package}
\alias{fdars}
\alias{fdars-package}
\title{fdars: Functional Data Analysis in 'Rust'}
\description{
Functional data analysis tools with a high-performance 'Rust' backend. Provides methods for functional data manipulation, depth computation, distance metrics, regression, and statistical testing. Supports both 1D functional data (curves) and 2D functional data (surfaces). Methods are described in Ramsay and Silverman (2005, ISBN:978-0-387-40080-8) "Functional Data Analysis" and Ferraty and Vieu (2006, ISBN:978-0-387-30369-7) "Nonparametric Functional Data Analysis".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sipemu.github.io/fdars-r/}
  \item \url{https://github.com/sipemu/fdars-r}
  \item Report bugs at \url{https://github.com/sipemu/fdars-r/issues}
}

}
\author{
\strong{Maintainer}: Simon Müller \email{simon.mueller@muon-stat.com}

}
\keyword{internal}
