% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{fdata.cen}
\alias{fdata.cen}
\title{Center functional data}
\usage{
fdata.cen(fdataobj)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}
}
\value{
A centered 'fdata' object.
}
\description{
Subtract the mean function from each curve.
}
\examples{
fd <- fdata(matrix(rnorm(100), 10, 10))
fd_centered <- fdata.cen(fd)
}
