% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{fdata2fd}
\alias{fdata2fd}
\title{Convert Functional Data to fd class}
\usage{
fdata2fd(fdataobj, nbasis = 10, type = c("bspline", "fourier"))
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{nbasis}{Number of basis functions (default 10).}

\item{type}{Type of basis: "bspline" (default) or "fourier".}
}
\value{
An object of class 'fd' from the fda package.
}
\description{
Converts an fdata object to an fd object from the fda package.
Requires the fda package to be installed.
}
\examples{
\dontshow{if (requireNamespace("fda", quietly = TRUE)) withAutoprint(\{ # examplesIf}
t <- seq(0, 1, length.out = 50)
X <- matrix(0, 20, 50)
for (i in 1:20) X[i, ] <- sin(2*pi*t) + rnorm(50, sd = 0.1)
fd <- fdata(X, argvals = t)
fd_obj <- fdata2fd(fd, nbasis = 10)
\dontshow{\}) # examplesIf}
}
