% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{fdata2pc}
\alias{fdata2pc}
\title{Convert Functional Data to Principal Component Scores}
\usage{
fdata2pc(fdataobj, ncomp = 2, lambda = 0, norm = TRUE)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{ncomp}{Number of principal components to extract (default 2).}

\item{lambda}{Regularization parameter (default 0, not currently used).}

\item{norm}{Logical. If TRUE (default), normalize the scores.}
}
\value{
A list with components:
\describe{
\item{d}{Singular values (proportional to sqrt of eigenvalues)}
\item{rotation}{fdata object containing PC loadings}
\item{x}{Matrix of PC scores (n x ncomp)}
\item{mean}{Mean function (numeric vector)}
\item{fdataobj.cen}{Centered fdata object}
\item{call}{The function call}
}
}
\description{
Performs functional PCA and returns principal component scores for
functional data. Uses SVD on centered data.
}
\examples{
t <- seq(0, 1, length.out = 50)
X <- matrix(0, 20, 50)
for (i in 1:20) X[i, ] <- sin(2*pi*t) + rnorm(50, sd = 0.1)
fd <- fdata(X, argvals = t)
pc <- fdata2pc(fd, ncomp = 3)
}
