% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{flm.test}
\alias{flm.test}
\title{Statistical Tests for Functional Data}
\usage{
flm.test(fdataobj, y, B = 500, ...)
}
\arguments{
\item{fdataobj}{An object of class 'fdata' (functional covariate).}

\item{y}{Response vector.}

\item{B}{Number of bootstrap samples for p-value computation.}

\item{...}{Additional arguments.}
}
\value{
A list of class 'htest' with components:
\describe{
\item{statistic}{The test statistic}
\item{p.value}{Bootstrap p-value}
\item{method}{Name of the test}
}
}
\description{
Functions for hypothesis testing with functional data.
Test for Functional Linear Model
}
\details{
Tests the goodness-of-fit for a functional linear model using
the projected Cramer-von Mises statistic.
}
\examples{
fd <- fdata(matrix(rnorm(200), 20, 10))
y <- rnorm(20)
\donttest{
test_result <- flm.test(fd, y, B = 100)
test_result$p.value
}
}
