% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{fmean.test.fdata}
\alias{fmean.test.fdata}
\title{Test for Equality of Functional Means}
\usage{
fmean.test.fdata(fdataobj, mu0 = NULL, B = 500, ...)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{mu0}{Hypothesized mean function (vector). If NULL, tests against zero.}

\item{B}{Number of bootstrap samples for p-value computation.}

\item{...}{Additional arguments.}
}
\value{
A list of class 'htest' with components:
\describe{
\item{statistic}{The test statistic}
\item{p.value}{Bootstrap p-value}
\item{method}{Name of the test}
}
}
\description{
Tests whether the mean function equals a specified value.
}
\examples{
fd <- fdata(matrix(rnorm(200), 20, 10))
\donttest{
test_result <- fmean.test.fdata(fd, B = 100)
test_result$p.value
}
}
