% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.R
\name{fregre.basis}
\alias{fregre.basis}
\title{Functional Basis Regression}
\usage{
fregre.basis(fdataobj, y, basis.x = NULL, basis.b = NULL, lambda = 0, ...)
}
\arguments{
\item{fdataobj}{An object of class 'fdata' (functional covariate).}

\item{y}{Response vector.}

\item{basis.x}{Basis for the functional covariate (currently ignored).}

\item{basis.b}{Basis for the coefficient function (currently ignored).}

\item{lambda}{Smoothing/regularization parameter (L2 penalty).}

\item{...}{Additional arguments.}
}
\value{
A fitted regression object of class 'fregre.fd' with components:
\item{coefficients}{Beta coefficient function values}
\item{intercept}{Intercept term}
\item{fitted.values}{Fitted values}
\item{residuals}{Residuals}
\item{lambda}{Regularization parameter used}
\item{r.squared}{R-squared (coefficient of determination)}
\item{mean.X}{Mean of functional covariate (for prediction)}
\item{mean.y}{Mean of response (for prediction)}
\item{sr2}{Residual variance}
\item{fdataobj}{Original functional data}
\item{y}{Response vector}
\item{call}{The function call}
}
\description{
Fits a functional linear model using basis expansion (ridge regression).
Uses the anofox-regression Rust backend for efficient L2-regularized regression.
}
