% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.R
\name{fregre.pc}
\alias{fregre.pc}
\title{Functional Regression}
\usage{
fregre.pc(fdataobj, y, ncomp = NULL, ...)
}
\arguments{
\item{fdataobj}{An object of class 'fdata' (functional covariate).}

\item{y}{Response vector.}

\item{ncomp}{Number of principal components to use.}

\item{...}{Additional arguments.}
}
\value{
A fitted regression object of class 'fregre.fd' with components:
\item{coefficients}{Beta coefficient function values}
\item{intercept}{Intercept term}
\item{fitted.values}{Fitted values}
\item{residuals}{Residuals}
\item{ncomp}{Number of components used}
\item{mean.X}{Mean of functional covariate (for prediction)}
\item{mean.y}{Mean of response (for prediction)}
\item{rotation}{PC loadings (for prediction)}
\item{l}{Indices of selected components}
\item{lm}{Underlying linear model}
\item{sr2}{Residual variance}
\item{fdataobj}{Original functional data}
\item{y}{Response vector}
\item{call}{The function call}
}
\description{
Functions for functional regression models.
Functional Principal Component Regression
}
\details{
Fits a functional linear model using principal component regression.
}
