% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{IKer.epa}
\alias{IKer.epa}
\title{Integrated Epanechnikov Kernel}
\usage{
IKer.epa(u)
}
\arguments{
\item{u}{Numeric vector of evaluation points.}
}
\value{
Cumulative integral values.
}
\description{
Integral of Ker.epa from -1 to u.
}
\examples{
u <- seq(-1.5, 1.5, length.out = 100)
plot(u, IKer.epa(u), type = "l", main = "Integrated Epanechnikov Kernel")
}
