% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{IKer.norm}
\alias{IKer.norm}
\title{Integrated Normal Kernel}
\usage{
IKer.norm(u)
}
\arguments{
\item{u}{Numeric vector of evaluation points.}
}
\value{
Cumulative integral of normal kernel from -Inf to u.
}
\description{
Integrated Normal Kernel
}
\examples{
u <- seq(-3, 3, length.out = 100)
plot(u, IKer.norm(u), type = "l", main = "Integrated Normal Kernel")
}
