% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{IKer.quar}
\alias{IKer.quar}
\title{Integrated Quartic Kernel}
\usage{
IKer.quar(u)
}
\arguments{
\item{u}{Numeric vector of evaluation points.}
}
\value{
Cumulative integral values.
}
\description{
Integral of Ker.quar from -1 to u.
}
\examples{
u <- seq(-1.5, 1.5, length.out = 100)
plot(u, IKer.quar(u), type = "l", main = "Integrated Quartic Kernel")
}
