% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{IKer.tri}
\alias{IKer.tri}
\title{Integrated Triweight Kernel}
\usage{
IKer.tri(u)
}
\arguments{
\item{u}{Numeric vector of evaluation points.}
}
\value{
Cumulative integral values.
}
\description{
Integral of Ker.tri from -1 to u.
}
\examples{
u <- seq(-1.5, 1.5, length.out = 100)
plot(u, IKer.tri(u), type = "l", main = "Integrated Triweight Kernel")
}
