% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{IKer.unif}
\alias{IKer.unif}
\title{Integrated Uniform Kernel}
\usage{
IKer.unif(u)
}
\arguments{
\item{u}{Numeric vector of evaluation points.}
}
\value{
Cumulative integral values.
}
\description{
Integral of Ker.unif from -1 to u.
}
\examples{
u <- seq(-1.5, 1.5, length.out = 100)
plot(u, IKer.unif(u), type = "l", main = "Integrated Uniform Kernel")
}
