% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irregFdata.R
\name{is.irregular}
\alias{is.irregular}
\title{Check if an Object is Irregular Functional Data}
\usage{
is.irregular(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
\code{TRUE} if \code{x} is of class \code{irregFdata}, \code{FALSE}
otherwise.
}
\description{
Check if an Object is Irregular Functional Data
}
\examples{
fd <- fdata(matrix(rnorm(100), 10, 10))
is.irregular(fd)  # FALSE

ifd <- sparsify(fd, minObs = 3, maxObs = 7)
is.irregular(ifd)  # TRUE
}
