% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{Ker.epa}
\alias{Ker.epa}
\title{Epanechnikov Kernel}
\usage{
Ker.epa(u)
}
\arguments{
\item{u}{Numeric vector of evaluation points.}
}
\value{
Kernel values at u (0 outside [-1, 1]).
}
\description{
Epanechnikov Kernel
}
\examples{
u <- seq(-1.5, 1.5, length.out = 100)
plot(u, Ker.epa(u), type = "l", main = "Epanechnikov Kernel")
}
