% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{Ker.tri}
\alias{Ker.tri}
\title{Triweight Kernel}
\usage{
Ker.tri(u)
}
\arguments{
\item{u}{Numeric vector of evaluation points.}
}
\value{
Kernel values at u (0 outside [-1, 1]).
}
\description{
Triweight Kernel
}
\examples{
u <- seq(-1.5, 1.5, length.out = 100)
plot(u, Ker.tri(u), type = "l", main = "Triweight Kernel")
}
