% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{kernel.add}
\alias{kernel.add}
\title{Add Covariance Functions}
\usage{
kernel.add(kernel1, kernel2)
}
\arguments{
\item{kernel1}{First covariance function.}

\item{kernel2}{Second covariance function.}
}
\value{
A combined covariance function.
}
\description{
Combines two covariance functions by addition.
}
\examples{
# Combine Gaussian with white noise
k_signal <- kernel.gaussian(variance = 1, length_scale = 0.2)
k_noise <- kernel.whitenoise(variance = 0.1)
k_total <- kernel.add(k_signal, k_noise)

t <- seq(0, 1, length.out = 50)
fd <- make.gaussian.process(n = 5, t = t, cov = k_total)
}
