% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{kernel.linear}
\alias{kernel.linear}
\title{Linear Covariance Function}
\usage{
kernel.linear(variance = 1, offset = 0)
}
\arguments{
\item{variance}{Variance parameter \eqn{\sigma^2} (default 1).}

\item{offset}{Offset parameter \eqn{c} (default 0).}
}
\value{
A covariance function object of class 'kernel_linear'.
}
\description{
Computes the linear covariance function:
\deqn{k(s, t) = \sigma^2 (s - c)(t - c)}
}
\details{
The linear covariance function produces sample paths that are linear
functions. It is useful when the underlying process is expected to
have a linear trend.
}
\examples{
# Generate linear function samples
cov_func <- kernel.linear(variance = 1)
t <- seq(0, 1, length.out = 50)
fd <- make.gaussian.process(n = 10, t = t, cov = cov_func)
plot(fd)
}
\seealso{
\code{\link{kernel.polynomial}}, \code{\link{make.gaussian.process}}
}
