% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{kernel.mult}
\alias{kernel.mult}
\title{Multiply Covariance Functions}
\usage{
kernel.mult(kernel1, kernel2)
}
\arguments{
\item{kernel1}{First covariance function.}

\item{kernel2}{Second covariance function.}
}
\value{
A combined covariance function.
}
\description{
Combines two covariance functions by multiplication.
}
\examples{
# Multiply periodic with Gaussian for locally periodic behavior
k_periodic <- kernel.periodic(period = 0.3)
k_gaussian <- kernel.gaussian(length_scale = 0.5)
k_prod <- kernel.mult(k_periodic, k_gaussian)
}
