% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{kernel.whitenoise}
\alias{kernel.whitenoise}
\title{White Noise Covariance Function}
\usage{
kernel.whitenoise(variance = 1)
}
\arguments{
\item{variance}{Variance (noise level) parameter \eqn{\sigma^2} (default 1).}
}
\value{
A covariance function object of class 'kernel_whitenoise'.
}
\description{
Computes the white noise covariance function:
\deqn{k(s, t) = \sigma^2 \mathbf{1}_{s = t}}
}
\details{
where \eqn{\mathbf{1}_{s = t}} is 1 if \eqn{s = t} and 0 otherwise.

The white noise covariance function represents independent noise at each
point. It can be added to other covariance functions to model observation
noise.
}
\examples{
# White noise covariance produces independent samples at each point
cov_func <- kernel.whitenoise(variance = 0.1)
t <- seq(0, 1, length.out = 50)
K <- cov_func(t)
# K is diagonal
}
\seealso{
\code{\link{kernel.gaussian}}, \code{\link{make.gaussian.process}}
}
