% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{median}
\alias{median}
\title{Compute Functional Median}
\usage{
median(
  fdataobj,
  method = c("FM", "mode", "RP", "RT", "BD", "MBD", "FSD", "KFSD", "RPD"),
  ...
)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{method}{Depth method to use. One of "FM", "mode", "RP", "RT", "BD",
"MBD", "FSD", "KFSD", or "RPD". Default is "FM".}

\item{...}{Additional arguments passed to the depth function.}
}
\value{
The curve (as fdata object) with maximum depth.
}
\description{
Returns the curve with maximum depth using the specified depth method.
}
\examples{
fd <- fdata(matrix(rnorm(100), 10, 10))
med <- median(fd)
med_mode <- median(fd, method = "mode")
}
