% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric.DTW}
\alias{metric.DTW}
\title{Dynamic Time Warping for Functional Data}
\usage{
metric.DTW(fdataobj, fdataref = NULL, p = 2, w = NULL, ...)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{fdataref}{An object of class 'fdata'. If NULL, computes self-distances.}

\item{p}{The p in Lp distance (default 2 for L2/Euclidean).}

\item{w}{Sakoe-Chiba window constraint. Default is min(ncol(fdataobj), ncol(fdataref)).
Use -1 for no window constraint.}

\item{...}{Additional arguments (ignored).}
}
\value{
A distance matrix.
}
\description{
Computes the Dynamic Time Warping distance between functional data.
DTW allows for non-linear alignment of curves.
}
\examples{
fd <- fdata(matrix(rnorm(100), 10, 10))
D <- metric.DTW(fd)
}
