% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric.hausdorff}
\alias{metric.hausdorff}
\title{Hausdorff Metric for Functional Data}
\usage{
metric.hausdorff(fdataobj, fdataref = NULL, ...)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{fdataref}{An object of class 'fdata'. If NULL, uses fdataobj.}

\item{...}{Additional arguments (ignored).}
}
\value{
A distance matrix.
}
\description{
Computes the Hausdorff distance between functional data objects.
The Hausdorff distance treats each curve as a set of points (t, f(t))
in 2D space and computes the maximum of the minimum distances.
}
\examples{
fd <- fdata(matrix(rnorm(100), 10, 10))
D <- metric.hausdorff(fd)
}
