% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{normalize}
\alias{normalize}
\title{Normalize functional data}
\usage{
normalize(fdataobj, p = 2)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{p}{The order of the norm (default 2 for L2 norm).}
}
\value{
A normalized 'fdata' object where each curve has unit norm.
}
\description{
Scales each curve to have Lp norm equal to 1.
}
\examples{
fd <- fdata(matrix(rnorm(100), 10, 10), argvals = seq(0, 1, length.out = 10))
fd_norm <- normalize(fd)
norm(fd_norm)  # All values should be 1
}
