% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{plot.amplitude_modulation}
\alias{plot.amplitude_modulation}
\title{Plot method for amplitude_modulation objects}
\usage{
\method{plot}{amplitude_modulation}(x, ...)
}
\arguments{
\item{x}{An amplitude_modulation object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Plot method for amplitude_modulation objects
}
