% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{plot.basis.cv}
\alias{plot.basis.cv}
\title{Plot method for basis.cv objects}
\usage{
\method{plot}{basis.cv}(x, ...)
}
\arguments{
\item{x}{A basis.cv object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{ggplot} object (invisibly).
}
\description{
Plot method for basis.cv objects
}
