% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot.cluster.fcm}
\alias{plot.cluster.fcm}
\title{Plot Method for cluster.fcm Objects}
\usage{
\method{plot}{cluster.fcm}(x, type = c("curves", "membership"), ...)
}
\arguments{
\item{x}{An object of class 'cluster.fcm'.}

\item{type}{Type of plot: "curves" (default) or "membership".}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A ggplot object.
}
\description{
Plot Method for cluster.fcm Objects
}
