% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot.cluster.kmeans}
\alias{plot.cluster.kmeans}
\title{Plot Method for cluster.kmeans Objects}
\usage{
\method{plot}{cluster.kmeans}(x, ...)
}
\arguments{
\item{x}{An object of class 'cluster.kmeans'.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A ggplot object.
}
\description{
Plot Method for cluster.kmeans Objects
}
