% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot.cluster.optim}
\alias{plot.cluster.optim}
\title{Plot Method for cluster.optim Objects}
\usage{
\method{plot}{cluster.optim}(x, ...)
}
\arguments{
\item{x}{An object of class 'cluster.optim'.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A ggplot object.
}
\description{
Plot Method for cluster.optim Objects
}
