% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{plot.group.distance}
\alias{plot.group.distance}
\title{Plot method for group.distance}
\usage{
\method{plot}{group.distance}(x, type = c("heatmap", "dendrogram"), which = NULL, ...)
}
\arguments{
\item{x}{An object of class 'group.distance'.}

\item{type}{Plot type: "heatmap" or "dendrogram".}

\item{which}{Which distance matrix to plot. If NULL (default), uses the first
available matrix from the group.distance object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object (for heatmap) or NULL (for dendrogram, uses base graphics).
}
\description{
Plot method for group.distance
}
