% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{plot.magnitudeshape}
\alias{plot.magnitudeshape}
\title{Plot Method for magnitudeshape Objects}
\usage{
\method{plot}{magnitudeshape}(x, ...)
}
\arguments{
\item{x}{An object of class 'magnitudeshape'.}

\item{...}{Additional arguments (ignored).}
}
\value{
The ggplot object (invisibly).
}
\description{
Plot Method for magnitudeshape Objects
}
