% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{plot.outliers.fdata}
\alias{plot.outliers.fdata}
\title{Plot method for outliers.fdata objects}
\usage{
\method{plot}{outliers.fdata}(x, col.outliers = "red", ...)
}
\arguments{
\item{x}{An object of class 'outliers.fdata'.}

\item{col.outliers}{Color for outlier curves (default "red").}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A ggplot object.
}
\description{
Plot method for outliers.fdata objects
}
