% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{plot.pspline.2d}
\alias{plot.pspline.2d}
\title{Plot method for pspline.2d objects}
\usage{
\method{plot}{pspline.2d}(x, ...)
}
\arguments{
\item{x}{A pspline.2d object.}

\item{...}{Additional arguments passed to plot.fdata.}
}
\value{
A \code{ggplot} object (invisibly).
}
\description{
Plot method for pspline.2d objects
}
