% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{plot.pspline}
\alias{plot.pspline}
\title{Plot method for pspline objects}
\usage{
\method{plot}{pspline}(x, ...)
}
\arguments{
\item{x}{A pspline object.}

\item{...}{Additional arguments passed to plot.fdata.}
}
\value{
A \code{ggplot} object (invisibly).
}
\description{
Plot method for pspline objects
}
