% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{plot.register.fd}
\alias{plot.register.fd}
\title{Plot Method for register.fd Objects}
\usage{
\method{plot}{register.fd}(x, type = c("registered", "original", "both"), ...)
}
\arguments{
\item{x}{An object of class 'register.fd'.}

\item{type}{Type of plot: "registered" (default), "original", or "both".}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A ggplot object.
}
\description{
Plot Method for register.fd Objects
}
