% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{plot.ssa_result}
\alias{plot.ssa_result}
\title{Plot method for ssa_result objects}
\usage{
\method{plot}{ssa_result}(x, type = c("decomposition", "spectrum"), curves = 1, ...)
}
\arguments{
\item{x}{An ssa_result object.}

\item{type}{Plot type: "decomposition" or "spectrum".}

\item{curves}{Indices of curves to plot.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Plot method for ssa_result objects
}
