% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{plot.stl_result}
\alias{plot.stl_result}
\title{Plot method for stl_result objects}
\usage{
\method{plot}{stl_result}(x, curves = 1, ...)
}
\arguments{
\item{x}{An stl_result object.}

\item{curves}{Indices of curves to plot.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Plot method for stl_result objects
}
