% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pred.MAE}
\alias{pred.MAE}
\title{Mean Absolute Error}
\usage{
pred.MAE(y_true, y_pred)
}
\arguments{
\item{y_true}{Actual values.}

\item{y_pred}{Predicted values.}
}
\value{
The mean absolute error.
}
\description{
Compute the Mean Absolute Error between predicted and actual values.
}
\examples{
y_true <- c(1, 2, 3, 4, 5)
y_pred <- c(1.1, 2.2, 2.9, 4.1, 4.8)
pred.MAE(y_true, y_pred)
}
