% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pred.R2}
\alias{pred.R2}
\title{R-Squared (Coefficient of Determination)}
\usage{
pred.R2(y_true, y_pred)
}
\arguments{
\item{y_true}{Actual values.}

\item{y_pred}{Predicted values.}
}
\value{
The R-squared value.
}
\description{
Compute the R-squared value between predicted and actual values.
}
\examples{
y_true <- c(1, 2, 3, 4, 5)
y_pred <- c(1.1, 2.2, 2.9, 4.1, 4.8)
pred.R2(y_true, y_pred)
}
