% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.R
\name{predict.fregre.np.multi}
\alias{predict.fregre.np.multi}
\title{Predict method for fregre.np.multi}
\usage{
\method{predict}{fregre.np.multi}(object, newdata.list = NULL, ...)
}
\arguments{
\item{object}{Fitted fregre.np.multi object.}

\item{newdata.list}{List of new fdata objects (same length as original).}

\item{...}{Additional arguments.}
}
\value{
Predicted values.
}
\description{
Predict method for fregre.np.multi
}
