% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.amplitude_modulation}
\alias{print.amplitude_modulation}
\title{Print method for amplitude_modulation objects}
\usage{
\method{print}{amplitude_modulation}(x, ...)
}
\arguments{
\item{x}{An amplitude_modulation object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for amplitude_modulation objects
}
