% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{print.basis.auto}
\alias{print.basis.auto}
\title{Print method for basis.auto objects}
\usage{
\method{print}{basis.auto}(x, ...)
}
\arguments{
\item{x}{A basis.auto object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for basis.auto objects
}
