% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{print.basis.cv}
\alias{print.basis.cv}
\title{Print method for basis.cv objects}
\usage{
\method{print}{basis.cv}(x, ...)
}
\arguments{
\item{x}{A basis.cv object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for basis.cv objects
}
