% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.cfd_autoperiod_result}
\alias{print.cfd_autoperiod_result}
\title{Print method for cfd_autoperiod_result objects}
\usage{
\method{print}{cfd_autoperiod_result}(x, ...)
}
\arguments{
\item{x}{A cfd_autoperiod_result object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for cfd_autoperiod_result objects
}
