% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{print.cluster.kmeans}
\alias{print.cluster.kmeans}
\title{Print Method for cluster.kmeans Objects}
\usage{
\method{print}{cluster.kmeans}(x, ...)
}
\arguments{
\item{x}{An object of class 'cluster.kmeans'.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for cluster.kmeans Objects
}
