% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{print.cluster.optim}
\alias{print.cluster.optim}
\title{Print Method for cluster.optim Objects}
\usage{
\method{print}{cluster.optim}(x, ...)
}
\arguments{
\item{x}{An object of class 'cluster.optim'.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for cluster.optim Objects
}
