% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.decomposition}
\alias{print.decomposition}
\title{Print method for decomposition objects}
\usage{
\method{print}{decomposition}(x, ...)
}
\arguments{
\item{x}{A decomposition object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for decomposition objects
}
