% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{print.fbplot}
\alias{print.fbplot}
\title{Print Method for fbplot Objects}
\usage{
\method{print}{fbplot}(x, ...)
}
\arguments{
\item{x}{An object of class 'fbplot'.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for fbplot Objects
}
