% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{print.fdata2pc}
\alias{print.fdata2pc}
\title{Print Method for FPCA Results}
\usage{
\method{print}{fdata2pc}(x, ...)
}
\arguments{
\item{x}{An object of class 'fdata2pc'.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for FPCA Results
}
