% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{print.fequiv.test}
\alias{print.fequiv.test}
\title{Print method for fequiv.test}
\usage{
\method{print}{fequiv.test}(x, ...)
}
\arguments{
\item{x}{An \code{fequiv.test} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for fequiv.test
}
