% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.R
\name{print.fregre.fd}
\alias{print.fregre.fd}
\title{Print method for fregre objects}
\usage{
\method{print}{fregre.fd}(x, ...)
}
\arguments{
\item{x}{A fregre.fd object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for fregre objects
}
