% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.R
\name{print.fregre.np.multi}
\alias{print.fregre.np.multi}
\title{Print method for fregre.np.multi}
\usage{
\method{print}{fregre.np.multi}(x, ...)
}
\arguments{
\item{x}{A fregre.np.multi object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for fregre.np.multi
}
