% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.R
\name{print.fregre.np}
\alias{print.fregre.np}
\title{Print method for fregre.np objects}
\usage{
\method{print}{fregre.np}(x, ...)
}
\arguments{
\item{x}{A fregre.np object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for fregre.np objects
}
