% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{print.group.distance}
\alias{print.group.distance}
\title{Print method for group.distance}
\usage{
\method{print}{group.distance}(x, digits = 3, ...)
}
\arguments{
\item{x}{A group.distance object.}

\item{digits}{Number of digits for printing (default 3).}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for group.distance
}
