% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{print.group.test}
\alias{print.group.test}
\title{Print method for group.test}
\usage{
\method{print}{group.test}(x, ...)
}
\arguments{
\item{x}{A group.test object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for group.test
}
