% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irregFdata.R
\name{print.irregFdata}
\alias{print.irregFdata}
\title{Print method for irregFdata objects}
\usage{
\method{print}{irregFdata}(x, ...)
}
\arguments{
\item{x}{An irregFdata object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for irregFdata objects
}
