% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{print.kernel}
\alias{print.kernel}
\title{Print Method for Covariance Functions}
\usage{
\method{print}{kernel}(x, ...)
}
\arguments{
\item{x}{A covariance function object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for Covariance Functions
}
