% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.lomb_scargle_result}
\alias{print.lomb_scargle_result}
\title{Print method for lomb_scargle_result objects}
\usage{
\method{print}{lomb_scargle_result}(x, ...)
}
\arguments{
\item{x}{A lomb_scargle_result object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for lomb_scargle_result objects
}
