% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{print.magnitudeshape}
\alias{print.magnitudeshape}
\title{Print Method for magnitudeshape Objects}
\usage{
\method{print}{magnitudeshape}(x, ...)
}
\arguments{
\item{x}{An object of class 'magnitudeshape'.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for magnitudeshape Objects
}
