% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{print.multiFunData}
\alias{print.multiFunData}
\title{Print method for multiFunData objects}
\usage{
\method{print}{multiFunData}(x, ...)
}
\arguments{
\item{x}{A multiFunData object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for multiFunData objects
}
