% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.multiple_periods}
\alias{print.multiple_periods}
\title{Print method for multiple_periods objects}
\usage{
\method{print}{multiple_periods}(x, ...)
}
\arguments{
\item{x}{A multiple_periods object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for multiple_periods objects
}
