% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{print.outliergram}
\alias{print.outliergram}
\title{Print Method for Outliergram Objects}
\usage{
\method{print}{outliergram}(x, ...)
}
\arguments{
\item{x}{An object of class 'outliergram'.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for Outliergram Objects
}
