% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{print.outliers.fdata}
\alias{print.outliers.fdata}
\title{Print method for outliers.fdata objects}
\usage{
\method{print}{outliers.fdata}(x, ...)
}
\arguments{
\item{x}{An outliers.fdata object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for outliers.fdata objects
}
