% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.peak_detection}
\alias{print.peak_detection}
\title{Print method for peak_detection objects}
\usage{
\method{print}{peak_detection}(x, ...)
}
\arguments{
\item{x}{A peak_detection object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for peak_detection objects
}
