% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.period_estimate}
\alias{print.period_estimate}
\title{Print method for period_estimate objects}
\usage{
\method{print}{period_estimate}(x, ...)
}
\arguments{
\item{x}{A period_estimate object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for period_estimate objects
}
