% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{print.pspline.2d}
\alias{print.pspline.2d}
\title{Print method for pspline.2d objects}
\usage{
\method{print}{pspline.2d}(x, ...)
}
\arguments{
\item{x}{A pspline.2d object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for pspline.2d objects
}
