% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{print.pspline}
\alias{print.pspline}
\title{Print method for pspline objects}
\usage{
\method{print}{pspline}(x, ...)
}
\arguments{
\item{x}{A pspline object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for pspline objects
}
