% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{print.register.fd}
\alias{print.register.fd}
\title{Print Method for register.fd Objects}
\usage{
\method{print}{register.fd}(x, ...)
}
\arguments{
\item{x}{An object of class 'register.fd'.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for register.fd Objects
}
