% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.sazed_result}
\alias{print.sazed_result}
\title{Print method for sazed_result objects}
\usage{
\method{print}{sazed_result}(x, ...)
}
\arguments{
\item{x}{A sazed_result object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for sazed_result objects
}
