% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.seasonality_changes}
\alias{print.seasonality_changes}
\title{Print method for seasonality_changes objects}
\usage{
\method{print}{seasonality_changes}(x, ...)
}
\arguments{
\item{x}{A seasonality_changes object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for seasonality_changes objects
}
