% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.seasonality_changes_auto}
\alias{print.seasonality_changes_auto}
\title{Print method for seasonality_changes_auto objects}
\usage{
\method{print}{seasonality_changes_auto}(x, ...)
}
\arguments{
\item{x}{A seasonality_changes_auto object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for seasonality_changes_auto objects
}
