% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.seasonality_classification}
\alias{print.seasonality_classification}
\title{Print method for seasonality_classification objects}
\usage{
\method{print}{seasonality_classification}(x, ...)
}
\arguments{
\item{x}{A seasonality_classification object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for seasonality_classification objects
}
