% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.ssa_result}
\alias{print.ssa_result}
\title{Print method for ssa_result objects}
\usage{
\method{print}{ssa_result}(x, ...)
}
\arguments{
\item{x}{An ssa_result object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for ssa_result objects
}
