% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{print.stl_result}
\alias{print.stl_result}
\title{Print method for stl_result objects}
\usage{
\method{print}{stl_result}(x, ...)
}
\arguments{
\item{x}{An stl_result object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for stl_result objects
}
