% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{semimetric.deriv}
\alias{semimetric.deriv}
\title{Semi-metric based on Derivatives}
\usage{
semimetric.deriv(fdataobj, fdataref = NULL, nderiv = 1, lp = 2, ...)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{fdataref}{An object of class 'fdata'. If NULL, uses fdataobj.}

\item{nderiv}{Derivative order (1, 2, ...). Default is 1.}

\item{lp}{The p in Lp metric. Default is 2 (L2 distance).}

\item{...}{Additional arguments passed to deriv.}
}
\value{
A distance matrix based on derivative distances.
}
\description{
Computes a semi-metric based on the Lp distance of the nderiv-th derivative
of functional data.
}
\examples{
# Create smooth curves
t <- seq(0, 2*pi, length.out = 100)
X <- matrix(0, 10, 100)
for (i in 1:10) X[i, ] <- sin(t + i/5)
fd <- fdata(X, argvals = t)

# Compute distance based on first derivative
D <- semimetric.deriv(fd, nderiv = 1)
}
