% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{semimetric.pca}
\alias{semimetric.pca}
\title{Semi-metric based on Principal Components}
\usage{
semimetric.pca(fdataobj, fdataref = NULL, ncomp = 2, ...)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{fdataref}{An object of class 'fdata'. If NULL, uses fdataobj.}

\item{ncomp}{Number of principal components to use.}

\item{...}{Additional arguments (ignored).}
}
\value{
A distance matrix based on PC scores.
}
\description{
Computes a semi-metric based on the first ncomp principal component scores.
}
\examples{
fd <- fdata(matrix(rnorm(200), 20, 10))
D <- semimetric.pca(fd, ncomp = 3)
}
