% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R, R/irregFdata.R
\name{standardize}
\alias{standardize}
\alias{standardize.fdata}
\alias{standardize.irregFdata}
\title{Standardize functional data (z-score normalization)}
\usage{
standardize(fdataobj)

\method{standardize}{fdata}(fdataobj)

\method{standardize}{irregFdata}(fdataobj)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}
}
\value{
A standardized 'fdata' object where each curve has mean 0 and sd 1.
}
\description{
Transforms each curve to have mean 0 and standard deviation 1.
This is useful for comparing curve shapes regardless of their level or scale.
}
\examples{
fd <- fdata(matrix(rnorm(100) * 10 + 50, 10, 10), argvals = seq(0, 1, length.out = 10))
fd_std <- standardize(fd)
# Check: each curve now has mean ~0 and sd ~1
rowMeans(fd_std$data)
apply(fd_std$data, 1, sd)
}
