% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{[.fdata}
\alias{[.fdata}
\title{Subset method for fdata objects}
\usage{
\method{[}{fdata}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{An object of class 'fdata'.}

\item{i}{Row indices (which curves to keep).}

\item{j}{Column indices (which time points to keep).}

\item{drop}{Logical. If TRUE and only one curve selected, return vector.}
}
\value{
An \code{fdata} object containing the selected subset.
}
\description{
Subset method for fdata objects
}
