% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irregFdata.R
\name{[.irregFdata}
\alias{[.irregFdata}
\title{Subset method for irregFdata objects}
\usage{
\method{[}{irregFdata}(x, i, ...)
}
\arguments{
\item{x}{An irregFdata object.}

\item{i}{Indices of observations to select.}

\item{...}{Additional arguments (ignored).}
}
\value{
An \code{irregFdata} object containing the selected subset.
}
\description{
Subset method for irregFdata objects
}
