% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irregFdata.R
\name{summary.irregFdata}
\alias{summary.irregFdata}
\title{Summary method for irregFdata objects}
\usage{
\method{summary}{irregFdata}(object, ...)
}
\arguments{
\item{object}{An irregFdata object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Summary method for irregFdata objects
}
