% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{var}
\alias{var}
\title{Functional Variance}
\usage{
var(fdataobj, ...)
}
\arguments{
\item{fdataobj}{An object of class 'fdata'.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
An fdata object containing the variance function (1D or 2D).
}
\description{
Computes the pointwise variance function of functional data.
}
\examples{
# 1D functional data
fd <- fdata(matrix(rnorm(100), 10, 10))
v <- var(fd)

# 2D functional data
X <- array(rnorm(500), dim = c(5, 10, 10))
fd2d <- fdata(X, argvals = list(1:10, 1:10), fdata2d = TRUE)
v2d <- var(fd2d)
}
